add catalog eclipse topology validation
